+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Next
 
 - **[Breaking change]** Update `DoAbc` to allow optional flags and name.
+- **[Breaking change]** Rename custom image media type to be closer to the corresponding tag names.
 
 # 0.10.0 (2019-11-21)